# docs/openapi.yaml
openapi: 3.0.3
info:
  title: go-tari-explorer API
  description: >-
    Read-only JSON API over this (L1) Tari block explorer's Postgres-backed data:
    indexed blocks (with per-kernel/per-output breakdown), pool-wide mining statistics,
    height-bucketed historical analysis (algo distribution, pool share, per-pool algo
    breakdown, block time, difficulty, rewards by algo/pool), a live chain-tip snapshot,
    and a liveness probe. Every route in this namespace is served by
    `internal/server/api.go`, mirrors one existing HTML route's Store/DB query and
    pagination/filter semantics 1:1 (see that file's own doc comments), and is strictly
    additive alongside (never a replacement for) this repo's existing server-rendered
    HTML+HTMX UI.

    This covers only the Phase 1 `/api/*` surface (blocks/pool-stats/analysis/tip-info/
    health) that exists as of this spec's writing. A Phase 2 surface (search/tx-state/
    mempool - the live-GRPC, rate-limited routes) is a separate, not-yet-written follow-up
    and is intentionally not documented here.
  version: "0.1.0"
  contact:
    name: go-tari-explorer
    url: https://github.com/Snipa22/go-tari-explorer
servers:
  # No production deployment URL for this repo is documented anywhere in its own
  # README.md/AGENTS.md/CONTRIBUTING.md, nor discoverable from any Dockerfile/systemd
  # unit/CI deploy step (checked before writing this - this repo has none of those).
...
  - url: http://localhost:8080
...
paths:
  /api/blocks:
    get:
      operationId: listBlocks
      tags: [blocks]
      summary: List recently indexed blocks
      description: >-
        Same before-cursor + limit pagination shape as the HTML `/` and `/blocks/partial`
        routes (`s.DB.ListBlocks`), collapsed into one JSON endpoint (no separate
        "full page vs HTMX rows-only partial" split). Rows are ordered by height DESC
        (most recent first).
      parameters:
        - name: before
          in: query
          required: false
          description: >-
            Height cursor - returns rows with `height < before`. Defaults to the
            maximum int64 value (i.e. the first/most-recent page) when omitted. A
            value that fails to parse as an int64 is rejected with a 400.
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          description: >-
            Maximum number of rows to return. Defaults to 25 (PageSize) when omitted.
            Values above 100 (MaxAPILimit) are silently clamped down to 100; a value
            that is not a positive integer is rejected with a 400.
          schema:
            type: integer
            minimum: 1
            default: 25
            maximum: 100
      responses:
        "200":
          description: >-
            An array of blocks. An empty result is always marshaled as a JSON array
            `[]`, never `null`.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"
        "400":
          description: Malformed `before` or `limit` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Store/database failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/tip-info:
    get:
      operationId: getTipInfo
      tags: [tip-info]
      summary: Live base-node chain-tip snapshot
      description: >-
        A real, live `s.Node.GetTipInfo(ctx)` GRPC call against the configured
        base-node host(s) - unlike every other route in this namespace, this is not
        Postgres-backed. `best_block_hash`/`accumulated_difficulty` are hex-encoded
        (matching this repo's own hash-rendering convention elsewhere, e.g.
        `db.Block.Hash`/`PrevHash`), and `base_node_state` is the human enum name
        (e.g. `"LISTENING"`), not its raw wire int32.
      responses:
        "200":
          description: Current chain-tip snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTipInfo"
        "503":
          description: >-
            No base-node GRPC host configured (`s.Node` is nil), or the live GetTipInfo
            call failed (e.g. every configured host unreachable).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
...
  schemas:
    ApiError:
      type: object
      description: >-
        Shared error-response shape for every /api/* route's 400/404/500/503
        responses - this repo returns valid JSON even on error, never a plain-text
        body (unlike the HTML routes' `http.Error`).
      required: [error]
      properties:
        error:
          type: string
          example: failed to load blocks
    Block:
      type: object
      description: >-
        Mirrors `db.Block` exactly (marshaled directly, not through the HTML-only
        `blockView` adapter) - a full decomposition of the real Tari protobuf
        BlockHeader message plus its nested ProofOfWork message (flattened with a
        `pow_` prefix), plus block-level summary fields derived from the block body.
        `hash`/`prev_hash` are hex-encoded strings; every other `[]byte` field below
        marshals via `encoding/json`'s default `[]byte` behavior (a base64 string) -
        these are secondary merkle-root/offset/proof-of-work fields, not primary
        searchable identifiers, so they are NOT hex-encoded (unlike `/api/tip-info`'s
        DTO, which hex-encodes its two primary hash fields by contrast).
      required:
        - height
        - hash
        - version
        - prev_hash
        - timestamp
        - output_mr
        - block_output_mr
        - kernel_mr
        - input_mr
        - total_kernel_offset
        - nonce
        - kernel_mmr_size
        - output_mmr_size
        - total_script_offset
        - validator_node_mr
        - validator_node_size
        - pow_algo_raw
        - pow_data
        - pow_algo
        - difficulty
        - kernel_count
        - output_count
        - pool_tag
        - reward_micro_minotari
      properties:
        height:
          type: integer
          format: int64
          minimum: 0
        hash:
          type: string
          description: Hex-encoded BlockHeader.Hash.
        version:
          type: integer
          format: int32
          minimum: 0
        prev_hash:
          type: string
          description: Hex-encoded BlockHeader.PrevHash.
        timestamp:
          type: integer
          format: int64
          description: Unix seconds.
        output_mr:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.OutputMr).
        block_output_mr:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.BlockOutputMr).
        kernel_mr:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.KernelMr).
        input_mr:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.InputMr).
        total_kernel_offset:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.TotalKernelOffset).
        nonce:
          type: integer
          format: int64
          minimum: 0
        kernel_mmr_size:
          type: integer
          format: int64
          minimum: 0
        output_mmr_size:
          type: integer
          format: int64
          minimum: 0
        total_script_offset:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.TotalScriptOffset).
        validator_node_mr:
          type: string
          format: byte
          description: Base64-encoded raw bytes (BlockHeader.ValidatorNodeMr).
        validator_node_size:
          type: integer
          format: int64
          minimum: 0
        pow_algo_raw:
          type: integer
          format: int64
          minimum: 0
          description: >-
            Raw ProofOfWork.pow_algo wire id (0=RXM, 1=SHA3X, 2=RXT, 3=C29) - the
            unclassified value pow_algo below is derived from.
        pow_data:
          type: string
          format: byte
          description: Base64-encoded raw bytes (ProofOfWork.pow_data).
        pow_algo:
          type: string
          enum: [RXM, RXT, C29, SHA3X]
          description: Classified algo string (see internal/poolattr.AlgoFromRaw).
        difficulty:
          type: integer
          format: int64
        kernel_count:
          type: integer
          format: int32
          minimum: 0
        output_count:
          type: integer
          format: int32
          minimum: 0
        pool_tag:
          type: string
          nullable: true
          description: Null means unattributed.
        reward_micro_minotari:
          type: integer
          format: int64
          minimum: 0
          description: >-
            Raw MicroMinotari coinbase reward. 0 means "unknown" (BulletProofPlus-
            hidden, or not yet backfilled - see migrations/0008_reward_micro_minotari
            .up.sql), not "no reward".
...
    ApiTipInfo:
      type: object
      description: >-
        Mirrors the unexported `apiTipInfo` DTO, built by hand from the real
        `tari_generated.TipInfoResponse`/`MetaData` GRPC response fields (verified
        against this module's actual go-tari-grpc-lib/v3 dependency, not guessed).
      required:
        - best_block_height
        - best_block_hash
        - accumulated_difficulty
        - pruned_height
        - timestamp
        - initial_sync_achieved
        - base_node_state
        - failed_checkpoints
      properties:
        best_block_height:
          type: integer
          format: int64
          minimum: 0
        best_block_hash:
          type: string
          description: Hex-encoded MetaData.BestBlockHash.
        accumulated_difficulty:
          type: string
          description: Hex-encoded MetaData.AccumulatedDifficulty.
        pruned_height:
          type: integer
          format: int64
          minimum: 0
        timestamp:
          type: integer
          format: int64
          description: MetaData.Timestamp, unix seconds.
        initial_sync_achieved:
          type: boolean
        base_node_state:
          type: string
          enum:
            - START_UP
            - HEADER_SYNC
            - HORIZON_SYNC
            - CONNECTING
            - BLOCK_SYNC
            - LISTENING
            - SYNC_FAILED
          description: >-
            Human enum name (BaseNodeState.String()) rather than the raw int32 wire
            value.
        failed_checkpoints:
          type: boolean
...